Do not schedule RPCs known to be NSRE'd for retry when they are first issued.
When issuing an RPC that is destined to a region that is known to be
unavailable, we used to add a callback on the RPC to potentially retry
it.  Once the region appeared to be available again, we re-dispatched
the RPC and added another callback to potentially retry it.  This
resulted in the RPC's callback chain having two (or more) callbacks to
retry the RPC, which could cause the RPC to execute more than once.

This fixes #81.

Signed-off-by: Benoit Sigoure <tsunanet@gmail.com>