Fail the RPC when we fail to de-serialize its response.
Whenever we fail to de-serialize the response to an RPC, we throw an
exception that is caught by Netty, which then causes us to go into
exceptionCaught(), from which we ought to close the connection to HBase
and cleanup() all outstanding RPCs (either by re-scheduling them if we
can, or by making them fail with a ConnectionResetException).

The bug here is that the RPC for which we failed to de-serialize the
response is no longer considered "in flight" because we were handling
its response, so we never triggered its callback chain.  This change
fixes that by propagating the exception to the caller.