Fix compatibility issue with CDH5.
In the latest version of CDH5 (0.96.1.1+cdh5.0.2), the asynchbase client
fails if closing a scanner while there are still rows to be read.
Specifying numberOfRows == 0 avoids the issue (this matches the behavior
in the HBase synchronous client).

Without this fix, the client fails processing the close response with
the error: 'org.hbase.async.InvalidResponseException: Should not have
gotten any cell blocks, yet there are [num] bytes that follow the
protobuf response.  This should never happen.  Are you using an
incompatible version of HBase?'