Fix the hanging bug after a restart of HBase region servers.
The bug existed with HBase versions lower than 0.95. The bug was that
ProtocolVersionCB of RegionClient did not properly handle exceptions like
ServerNotRunningYetException during a HBase restart. The side effect was that
neither the becomeReady nor the cleanup method of the corresponding region
client was called because the channel was already connected and healthy. As a
result, any attempt of sending requests to the region server via sendRpc ended
up accumulating the requests to the pending_rpcs. The fix was closing the
corresponding channel when ProtocolVersionCB gets an exception.

Signed-off-by: Benoit Sigoure <tsunanet@gmail.com>