Fix off-by-one error in error handling of batched edits.
The code AND the comment were both wrong, when we `continue' at that
point, `n' doesn't get incremented, so we were looping one extra time,
which was leading to an uncaught IndexOutOfBoundsException.

This bug would typically occur when a region moves from one server to
another, or becomes otherwise temporarily unavailable (e.g. the memstore
is really full and the region is getting flushed and is momentarily
throwing RegionTooBusyException).

This closes #77.