Craft a very unlikely key when sending probe "exists" RPCs.
Sometimes we need to "poke" HBase to see if a region is online or
a table exists.  Given a key, we prepend some unique suffix to make
it a lot less likely that we hit a real key with our probe, as doing
so might have some implications on the RegionServer's memory usage.

This closes #82.